+++ src/main/java/a/a/a/c/f/a/c/a/f/AHEM.java
@@ -1,23 +1,24 @@
 package a.a.a.c.f.a.c.a.f;
 
 import a.a.a.c.e.a.k.a.EXF;
-import java.math.BigDecimal;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import java.math.BigDecimal;
 
 @XmlAccessorType(XmlAccessType.FIELD)
 public class AHEM extends AHEL<BigDecimal> {
-   public AHEM() {
-      EXF.getInstance().ICQ();
-   }
+    public AHEM() {
+        EXF.getInstance().ICQ();
+    }
 
-   public AHEM(String var1, BigDecimal var2) {
-      super(var1, var2);
-      EXF.getInstance().ICO();
-      EXF.getInstance().ICP();
-   }
+    public AHEM(String var1, BigDecimal var2) {
+        super(var1, var2);
+        EXF.getInstance().ICO();
+        EXF.getInstance().ICP();
+    }
 
-   public String getValueAsString() {
-      return this.AHXA != null ? ((BigDecimal)this.AHXA).toPlainString() : "NULL";
-   }
+    public String getValueAsString() {
+        return this.AHXA != null ? this.AHXA.toPlainString() : "NULL";
+    }
 }
