+++ src/main/java/a/a/a/c/f/a/c/a/f/AHEP.java
@@ -1,22 +1,23 @@
 package a.a.a.c.f.a.c.a.f;
 
 import a.a.a.c.e.a.k.a.EXF;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 
 @XmlAccessorType(XmlAccessType.FIELD)
 public class AHEP extends AHEL<Integer> {
-   public AHEP() {
-      EXF.getInstance().ICQ();
-   }
+    public AHEP() {
+        EXF.getInstance().ICQ();
+    }
 
-   public AHEP(String var1, Integer var2) {
-      super(var1, var2);
-      EXF.getInstance().ICO();
-      EXF.getInstance().ICP();
-   }
+    public AHEP(String var1, Integer var2) {
+        super(var1, var2);
+        EXF.getInstance().ICO();
+        EXF.getInstance().ICP();
+    }
 
-   public String getValueAsString() {
-      return this.AHXA != null ? ((Integer)this.AHXA).toString() : "NULL";
-   }
+    public String getValueAsString() {
+        return this.AHXA != null ? this.AHXA.toString() : "NULL";
+    }
 }
